cook/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.1
channels-redis==2.4.2
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.6
django-allauth==0.50.0
django-bootstrap-form==3.4
django-channels==0.7.0
django-ckeditor==6.1.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.12.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==0.6.2
oauthlib==3.2.0
packaging==21.3
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.1
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.0
zope.interface==5.4.0
